pyproject.toml
[tool.poetry]
name = "template-download-file-large"
version = "0.1.0"
description = ""
authors = ["nguyennt65"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.6"
dramatiq = "^1.14.2"
pika = "^1.3.1"
django-minio-backend = "^3.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
